Add files via upload
This section describes the implementation of a basic custom system call in the BeOS kernel. Kernel source code was modified and recompiled using the legacy build tools included in the BeOS developer kit. The system call was designed to return a predefined message to the user space. Testing was completed via a simple user-level application that invoked the new call. Functionality was confirmed by observing correct output from the user application.